In progress refactor
